WIP - retry-based django e2e testing
